modules/kafka_api/requirements.txt
kafka-python
SQLAlchemy==1.3.19
